+++ contracts/test/UpdateClient.t.sol
@@ -46,7 +46,7 @@ contract SP1ICS07UpdateClientTest is SP1ICS07TendermintTest {
     function test_ValidUpdateClient() public {
         FixtureTestCase[] memory testCases = fixtureTestCases();
 
-        for (uint i = 0; i < testCases.length; i++) {
+        for (uint256 i = 0; i < testCases.length; i++) {
             setUpTestWithFixture(testCases[i].fileName);
 
             // set a correct timestamp
